+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/RetryDeliveryTask.kt
@@ -5,12 +5,12 @@ import com.bugsnag.android.performance.Logger
 internal class RetryDeliveryTask(
     private val retryQueue: RetryQueue,
     private val delivery: Delivery,
-    private val connectivity: Connectivity
+    private val connectivity: Connectivity,
 ) : AbstractTask() {
     override fun execute(): Boolean {
         if (!connectivity.shouldAttemptDelivery()) {
             Logger.d("Skipping RetryDeliveryTask - no connectivity.")
-            return false;
+            return false
         }
 
         val nextPayload = retryQueue.next() ?: return false