+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanFactory.kt
@@ -2,7 +2,6 @@ package com.bugsnag.android.performance.internal
 
 import android.app.Activity
 import androidx.annotation.RestrictTo
-import com.bugsnag.android.performance.HasAttributes
 import com.bugsnag.android.performance.NetworkRequestInfo
 import com.bugsnag.android.performance.NetworkRequestInstrumentationCallback
 import com.bugsnag.android.performance.SpanContext
@@ -10,9 +9,10 @@ import com.bugsnag.android.performance.SpanKind
 import com.bugsnag.android.performance.SpanOptions
 import com.bugsnag.android.performance.ViewType
 import com.bugsnag.android.performance.internal.integration.NotifierIntegration
+import com.bugsnag.android.performance.internal.processing.AttributeLimits
 import java.util.UUID
 
-internal typealias AttributeSource = (target: HasAttributes) -> Unit
+internal typealias AttributeSource = (target: SpanImpl) -> Unit
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 public class SpanFactory(
@@ -20,6 +20,8 @@ public class SpanFactory(
     public val spanAttributeSource: AttributeSource = {},
 ) {
     public var networkRequestCallback: NetworkRequestInstrumentationCallback? = null
+    internal var attributeLimits: AttributeLimits? = null
+
     public fun createCustomSpan(
         name: String,
         options: SpanOptions = SpanOptions.DEFAULTS,
@@ -27,7 +29,7 @@ public class SpanFactory(
     ): SpanImpl {
         val isFirstClass = options.isFirstClass != false
         val span = createSpan(name, SpanKind.INTERNAL, SpanCategory.CUSTOM, options, spanProcessor)
-        span.setAttribute("bugsnag.span.first_class", isFirstClass)
+        span.attributes["bugsnag.span.first_class"] = isFirstClass
         return span
     }
 
@@ -48,8 +50,8 @@ public class SpanFactory(
                 options,
                 spanProcessor,
             )
-            span.setAttribute("http.url", resultUrl)
-            span.setAttribute("http.method", verbUpper)
+            span.attributes["http.url"] = resultUrl
+            span.attributes["http.method"] = verbUpper
             return span
         }
         return null
@@ -81,14 +83,14 @@ public class SpanFactory(
             spanProcessor,
         )
 
-        span.setAttribute("bugsnag.view.type", viewType.typeName)
-        span.setAttribute("bugsnag.view.name", viewName)
-        span.setAttribute("bugsnag.span.first_class", isFirstClass)
+        span.attributes["bugsnag.view.type"] = viewType.typeName
+        span.attributes["bugsnag.view.name"] = viewName
+        span.attributes["bugsnag.span.first_class"] = isFirstClass
 
         val appStart = SpanContext.contextStack.findSpan { it.category == SpanCategory.APP_START }
         if (appStart != null && appStart.attributes["bugsnag.app_start.first_view_name"] == null) {
-            appStart.setAttribute("bugsnag.view.type", viewType.typeName)
-            appStart.setAttribute("bugsnag.app_start.first_view_name", viewName)
+            appStart.attributes["bugsnag.view.type"] = viewType.typeName
+            appStart.attributes["bugsnag.app_start.first_view_name"] = viewName
         }
 
         return span
@@ -110,9 +112,9 @@ public class SpanFactory(
             spanProcessor,
         )
 
-        span.setAttribute("bugsnag.view.name", viewName)
-        span.setAttribute("bugsnag.view.type", viewType.typeName)
-        span.setAttribute("bugsnag.phase", phaseName)
+        span.attributes["bugsnag.view.name"] = viewName
+        span.attributes["bugsnag.view.type"] = viewType.typeName
+        span.attributes["bugsnag.phase"] = phaseName
 
         return span
     }
@@ -144,7 +146,7 @@ public class SpanFactory(
             spanProcessor,
         )
 
-        span.setAttribute("bugsnag.app_start.type", startType.lowercase())
+        span.attributes["bugsnag.app_start.type"] = startType.lowercase()
 
         return span
     }
@@ -162,7 +164,7 @@ public class SpanFactory(
             spanProcessor,
         )
 
-        span.setAttribute("bugsnag.phase", "FrameworkLoad")
+        span.attributes["bugsnag.phase"] = "FrameworkLoad"
 
         return span
     }
@@ -185,6 +187,7 @@ public class SpanFactory(
             parentSpanId = parentContext?.spanId ?: 0L,
             processor = spanProcessor,
             makeContext = options.makeContext,
+            attributeLimits = attributeLimits,
         )
 
         spanAttributeSource(span)