+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.7.0 (2024-09-23)
+
+### Bug fixes
+
+* Avoid crashing on Android 11 due to https://issuetracker.google.com/issues/175055271.
+  [#276](https://github.com/bugsnag/bugsnag-android-performance/pull/276)
+
+### Changes
+
+* Use API key subdomain as default Performance endpoint.
+  [#277](https://github.com/bugsnag/bugsnag-android-performance/pull/277)
+* Trace Propagation URLs and service name can be configured in `AndroidManifest.xml`
+  [#281](https://github.com/bugsnag/bugsnag-android-performance/pull/281)
+* Added configurable limit to number of span attributes per span.
+  [#280](https://github.com/bugsnag/bugsnag-android-performance/pull/280)
+
 ## 1.6.0 (2024-08-27)
 
 ### Enhancements